requirements.txt
django==3.2
tweepy==3.10.0
celery==4.1.1
redis==3.5.3
django-filter==2.4.0
python-dotenv==0.15.0
django-allauth